Made the bayesian fitting a package extra
